@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.DotNet.Scaffolding.Shared.Messaging;
 using SummitWear.Server.Models;
 using SummitWear.Server.Services;
 using System.Security.Claims;
@@ -25,7 +26,13 @@ namespace SummitWear.Server.Controllers
         {
             if (!ModelState.IsValid) return BadRequest(ModelState);
 
-            var user = new User { UserName = model.Email, Email = model.Email, FullName = model.FullName };
+            var user = new User
+            {
+                UserName = model.Email,  // Use Email as Username
+                Email = model.Email,
+                FullName = model.FullName
+            };
+
             var result = await _userManager.CreateAsync(user, model.Password);
 
             if (!result.Succeeded) return BadRequest(result.Errors);
@@ -34,9 +41,14 @@ namespace SummitWear.Server.Controllers
             return Ok(new { Message = "User registered successfully!" });
         }
 
+
         [HttpPost("login")]
         public async Task<IActionResult> Login([FromBody] Login model)
         {
+            if (model == null)
+            {
+                return BadRequest(new { Message = "invalid request" });
+            }
             var user = await _userManager.FindByEmailAsync(model.Email);
             if (user == null || !await _userManager.CheckPasswordAsync(user, model.Password))
             {
@@ -46,7 +58,7 @@ namespace SummitWear.Server.Controllers
             var roles = await _userManager.GetRolesAsync(user);
             var token = _jwtService.GenerateToken(user, (List<string>)roles); // JWT generation logic
 
-            return Ok(new { Token = token });
+            return Ok(new { Token = token, user.FullName });
         }
 
         [Authorize]
@@ -60,6 +72,28 @@ namespace SummitWear.Server.Controllers
             }
 
             return Ok(new { Message = "Valid token", UserId = userId });
+        }
+
+        [Authorize(Roles = "Admin")] // Ensure only Admins can access this
+        [HttpPost("promote-to-admin")]
+        public async Task<IActionResult> AssignRole([FromBody] AssignRoleDto model)
+        {
+            var user = await _userManager.FindByIdAsync(model.UserId);
+            if (user == null) return NotFound(new { Message = "User not found" });
+
+            if (!await _userManager.IsInRoleAsync(user, model.Role))
+            {
+                await _userManager.AddToRoleAsync(user, model.Role);
+                return Ok(new { Message = "Role assigned successfully!" });
+            }
+
+            return BadRequest(new { Message = "User already has this role." });
+        }
+
+        public class AssignRoleDto
+        {
+            public string UserId { get; set; }
+            public string Role { get; set; }
         }
     }
 }
