@@ -4,12 +4,13 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
-import { LoginComponent } from './features/login/pages/login/login.component';
+
+
 
 
 @NgModule({ declarations: [
         AppComponent,
-        LoginComponent,
+
     ],
   bootstrap: [
     AppComponent],
