@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-admin-manage',
-  templateUrl: './admin-manage.component.html',
-  styleUrl: './admin-manage.component.css'
+    selector: 'app-admin-manage',
+    templateUrl: './admin-manage.component.html',
+    styleUrl: './admin-manage.component.css',
+    standalone: false
 })
 export class AdminManageComponent {
 
