@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-analytics',
-  templateUrl: './analytics.component.html',
-  styleUrl: './analytics.component.css'
+    selector: 'app-analytics',
+    templateUrl: './analytics.component.html',
+    styleUrl: './analytics.component.css',
+    standalone: false
 })
 export class AnalyticsComponent {
 
