@@ -2,9 +2,10 @@ import { Component, AfterViewInit } from '@angular/core';
 import { Chart } from 'chart.js/auto';
 
 @Component({
-  selector: 'app-dashboard',
-  templateUrl: './dashboard.component.html',
-  styleUrl: './dashboard.component.css'
+    selector: 'app-dashboard',
+    templateUrl: './dashboard.component.html',
+    styleUrl: './dashboard.component.css',
+    standalone: false
 })
 export class DashboardComponent implements AfterViewInit {
   ngAfterViewInit() {
