@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-order-manage',
-  templateUrl: './order-manage.component.html',
-  styleUrl: './order-manage.component.css'
+    selector: 'app-order-manage',
+    templateUrl: './order-manage.component.html',
+    styleUrl: './order-manage.component.css',
+    standalone: false
 })
 export class OrderManageComponent {
   searchQuery: string = '';
