@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-about-us',
-  templateUrl: './about-us.component.html',
-  styleUrl: './about-us.component.css'
+    selector: 'app-about-us',
+    templateUrl: './about-us.component.html',
+    styleUrl: './about-us.component.css',
+    standalone: false
 })
 export class AboutUsComponent {
 
