@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-hero',
-  templateUrl: './hero.component.html',
-  styleUrl: './hero.component.css'
+    selector: 'app-hero',
+    templateUrl: './hero.component.html',
+    styleUrl: './hero.component.css',
+    standalone: false
 })
 export class HeroComponent {
 
