@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-our-promise',
-  templateUrl: './our-promise.component.html',
-  styleUrl: './our-promise.component.css'
+    selector: 'app-our-promise',
+    templateUrl: './our-promise.component.html',
+    styleUrl: './our-promise.component.css',
+    standalone: false
 })
 export class OurPromiseComponent {
 
