@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-promotion',
-  templateUrl: './promotion.component.html',
-  styleUrl: './promotion.component.css'
+    selector: 'app-promotion',
+    templateUrl: './promotion.component.html',
+    styleUrl: './promotion.component.css',
+    standalone: false
 })
 export class PromotionComponent {
 
