@@ -9,9 +9,10 @@ interface ClothingItem {
 }
 
 @Component({
-  selector: 'app-weather',
-  templateUrl: './weather.component.html',
-  styleUrls: ['./weather.component.css']
+    selector: 'app-weather',
+    templateUrl: './weather.component.html',
+    styleUrls: ['./weather.component.css'],
+    standalone: false
 })
 export class WeatherComponent implements OnInit {
   smallScreen: boolean = false; // Tracks if the screen size is small
