@@ -1,28 +1,60 @@
 import { Component } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Router } from '@angular/router';
+import { Router } from '@angular/router'
+import { AuthService } from '../../../core/services/auth.service';
 
 @Component({
-  selector: 'app-login',
-  templateUrl: './login.component.html',
-  styleUrl: './login.component.css'
+    selector: 'app-login',
+    templateUrl: './login.component.html',
+    styleUrl: './login.component.css',
+    standalone: false
 })
 export class LoginComponent {
   email = '';
   password = '';
+  confirmPassword = '';
+  isRegisterMode = false;
+  fullName = '';
 
-  constructor(private http: HttpClient, private router: Router) { }
+
+  constructor(private http: HttpClient, private router: Router, private authService: AuthService,) { }
+  toggleMode() {
+    this.isRegisterMode = !this.isRegisterMode;
+  }
 
   login() {
-    const loginData = { email: this.email, password: this.password };
+    if (!this.email || !this.password) {
+      console.error('Email or password is missing');
+      return;
+    }
 
-    this.http.post('https://your-api-url.com/api/auth/login', loginData)
-      .subscribe({
-        next: (response: any) => {
-          localStorage.setItem('token', response.token); // Store JWT
-          this.router.navigate(['/']); // Redirect after login
-        },
-        error: err => console.error('Login failed', err)
-      });
+    this.authService.login(this.email, this.password);
+    
   }
+
+  logout() {
+    this.authService.logout();
+  }
+
+  register() {
+    if (this.password !== this.confirmPassword) {
+      console.error('Passwords do not match!');
+      return;
+    }
+
+    const registerData = {
+      fullName: this.fullName,
+      email: this.email,
+      password: this.password
+    };
+
+    this.authService.register(registerData).subscribe({
+      next: () => {
+        console.log('Registration successful');
+        this.isRegisterMode = false; // Switch back to login
+      },
+      error: err => console.error('Registration failed:', err)
+    });
+  }
+
 }
