@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-privacy',
-  templateUrl: './privacy.component.html',
-  styleUrl: './privacy.component.css'
+    selector: 'app-privacy',
+    templateUrl: './privacy.component.html',
+    styleUrl: './privacy.component.css',
+    standalone: false
 })
 export class PrivacyComponent {
 
