@@ -1,10 +1,12 @@
 
 
 export interface Product {
-  productId: string;
-  name: string;
-  description: string;
-  price: number;
-  stock: number;
-  categoryId: string; // Foreign key for Category
+    id: number,
+    name: string,
+    price: number,
+    stock: number,
+    images: [],   
+    categoryId: number,
+    categorySeason: string,
+    categoryType: string, // Foreign key for Category
 }
